% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{control_extractor}
\alias{control_extractor}
\title{Extracts the control variable names and coefficients from an lm model
summary.}
\usage{
control_extractor(model, x, feols_model = FALSE)
}
\arguments{
\item{model}{A model summary object.}

\item{x}{A string containing the independent variable name.}

\item{feols_model}{An indicator for whether `model` is a `fixest::feols()`
model. Defaults to `FALSE`.}
}
\value{
A dataframe with two columns, `term` contains the name of the control
        and `coef` contains the coefficient estimate.
}
\description{
Extracts the control variable names and coefficients from a model summary.
}
\examples{
m <- summary(lm(Salnty ~ STheta + T_degC, bottles))
control_extractor(model = m, x = "STheta");

m <- summary(lm(Salnty ~ STheta*T_degC + O2Sat, bottles))
control_extractor(model = m, x = "STheta");
}
