% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{paste_factory}
\alias{paste_factory}
\title{Paste together controls and independent variable}
\usage{
paste_factory(controls, x)
}
\arguments{
\item{controls}{A vector of strings containing control variable names.}

\item{x}{A string containing the independent variable name.}
}
\value{
A string concatenating independent and control variables separated
         by '+'.
}
\description{
`paste_factory()` constructs the right hand side of the regression as a
a string i.e. "x + control1 + control2".
}
\examples{
paste_factory(controls = c("control1", "control2"),
              x = "independentVariable");
}
\keyword{internal}
