% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_aic}
\alias{plot_aic}
\title{Plots the AIC across model specifications.}
\usage{
plot_aic(sca_data, title = "", show_index = TRUE, plot_vars = TRUE)
}
\arguments{
\item{sca_data}{A data frame returned by `sca()` containing model estimates
from the specification curve analysis.}

\item{title}{A string to use as the plot title. Defaults to an empty string,
`""`.}

\item{show_index}{A boolean indicating whether to label the model index on the
the x-axis. Defaults to `TRUE`.}

\item{plot_vars}{A boolean indicating whether to include a panel on the plot
showing which variables are present in each model. Defaults
to `TRUE`.}
}
\value{
If `plot_vars = TRUE` a `patchwork` object combining the plot and the
        variable panel; if `plot_vars = FALSE` a ggplot object.
}
\description{
plot_aic() plots the Akaike information criterion across model specifications.
Only available for nonlinear regression models.
}
\examples{
plot_aic(sca_data = sca(y = "Salnty", x = "T_degC",
                       controls = c("ChlorA", "O2Sat"),
                       data = bottles, progress_bar = TRUE, parallel = FALSE),
                     title = "AIC");
plot_aic(sca_data = sca(y = "Salnty", x = "T_degC",
                       controls = c("ChlorA*O2Sat"),
                       data = bottles, progress_bar = FALSE,
                       parallel = FALSE),
                       show_index = FALSE, plot_vars = FALSE);
\donttest{
plot_aic(sca_data = sca(y = "Salnty", x = "T_degC",
                         controls = c("ChlorA*NO3uM", "O2Sat*NO3uM"),
                         data = bottles,
                         progress_bar = TRUE, parallel = TRUE, workers = 2));
}
}
