% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_r2_adj}
\alias{plot_r2_adj}
\title{Plots the adj. R-squared across model specifications.}
\usage{
plot_r2_adj(sca_data, title = "", show_index = TRUE, plot_vars = TRUE)
}
\arguments{
\item{sca_data}{A data frame returned by `sca()` containing model estimates
from the specification curve analysis.}

\item{title}{A string to use as the plot title. Defaults to an empty string,
`""`.}

\item{show_index}{A boolean indicating whether to label the model index on the
the x-axis. Defaults to `TRUE`.}

\item{plot_vars}{A boolean indicating whether to include a panel on the plot
showing which variables are present in each model. Defaults
to `TRUE`.}
}
\value{
If `plot_vars = TRUE` a `patchwork` object combining the plot and the
        variable panel; if `plot_vars = FALSE` a ggplot object.
}
\description{
plot_r2_adj() plots the adjusted R-squared across model specifications. Only
available for linear regression models. Note when fixed effects are
are specified the within adjusted R-squared is used (i.e. `fixest::r2()`
with `type="war2"`).
}
\examples{
plot_r2_adj(sca_data = sca(y = "Salnty", x = "T_degC",
                         controls = c("ChlorA", "O2Sat"),
                         data = bottles, progress_bar = TRUE,
                         parallel = FALSE),
                     title = "Adjusted R^2");
plot_r2_adj(sca_data = sca(y="Salnty", x="T_degC",
                         controls = c("ChlorA*O2Sat"),
                         data = bottles, progress_bar = FALSE,
                         parallel = FALSE),
                     show_index = FALSE, plot_vars = FALSE);
\donttest{
plot_r2_adj(sca_data = sca(y = "Salnty", x = "T_degC",
                         controls = c("ChlorA*NO3uM", "O2Sat*NO3uM"),
                         data = bottles,
                         progress_bar = TRUE, parallel = TRUE, workers = 2));
}
}
