% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_vars}
\alias{plot_vars}
\title{Plots the variables in each model.}
\usage{
plot_vars(sca_data, title = "", color_controls = FALSE)
}
\arguments{
\item{sca_data}{A data frame returned by `sca()` containing model estimates
from the specification curve analysis.}

\item{title}{A string to use as the plot title. Defaults to an empty string,
`""`.}

\item{color_controls}{A boolean indicating whether to give each variable a
color to improve readability. Defaults to `FALSE`.}
}
\value{
A ggplot object.
}
\description{
plot_vars() plots the variables included in each model specification in order
of model index. Returns a ggplot object that can then be combined with the
output of other functions like plot_rmse() if further customization of each
plot is desired.
}
\examples{
plot_vars(sca_data = sca(y = "Salnty", x = "T_degC",
                        controls = c("ChlorA", "O2Sat"),
                        data = bottles, progress_bar = TRUE,
                        parallel = FALSE),
                     title = "Model Variable Specifications");
plot_vars(sca_data = sca(y = "Salnty", x = "T_degC",
                        controls = c("ChlorA*O2Sat"),
                        data = bottles, progress_bar = FALSE,
                        parallel = FALSE),
                     color_controls = TRUE);
\donttest{
plot_vars(sca_data = sca(y = "Salnty", x = "T_degC",
                        controls = c("ChlorA*NO3uM", "O2Sat*NO3uM"),
                        data = bottles,
                        progress_bar = TRUE, parallel = TRUE, workers = 2));
}
}
