% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sca_test.R
\name{print.sca_test}
\alias{print.sca_test}
\title{Print a specification curve joint-inference test}
\usage{
\method{print}{sca_test}(x, ...)
}
\arguments{
\item{x}{An object of class `"sca_test"` returned by [sca_test()].}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print a specification curve joint-inference test
}
