% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{un_as_is}
\alias{un_as_is}
\title{Removes the `AsIs` class attribute from the input.}
\usage{
un_as_is(x)
}
\arguments{
\item{x}{An object with the `AsIs` class attribute.}
}
\value{
An object without the `AsIs` class attribute.
}
\description{
Removes the `AsIs` class attribute from the input. Taken from:
<https://stackoverflow.com/a/12866609>
}
\examples{
un_as_is(x = I(c(1:4)));
}
\keyword{internal}
