% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{duplicate_remover}
\alias{duplicate_remover}
\title{Removes duplicate control variables from user input}
\usage{
duplicate_remover(controls, x)
}
\arguments{
\item{controls}{A vector of strings containing control variable names.}

\item{x}{A string containing the independent variable name.}
}
\value{
A vector of strings containing control variable names
}
\description{
Removes duplicate control variables from user input
}
\examples{
duplicate_remover(controls = c("control1", "control2*control3"),
                  x = "independentVariable");
}
