% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotR2Adj}
\alias{plotR2Adj}
\title{Plots the adj. R-squared across model specifications.}
\usage{
plotR2Adj(sca_data, title = "", showIndex = TRUE, plotVars = TRUE)
}
\arguments{
\item{sca_data}{A data frame returned by `sca()` containing model estimates
from the specification curve analysis.}

\item{title}{A string to use as the plot title. Defaults to an empty string,
`""`.}

\item{showIndex}{A boolean indicating whether to label the model index on the
the x-axis. Defaults to `TRUE`.}

\item{plotVars}{A boolean indicating whether to include a panel on the plot
showing which variables are present in each model. Defaults
to `TRUE`.}
}
\value{
If `plotVars = TRUE` a `patchwork` object combining the plot and the
        variable panel; if `plotVars = FALSE` a ggplot object.
}
\description{
plotR2Adj() plots the adjusted R-squared across model specifications. Only
available for linear regression models. Note when fixed effects are
are specified the within adjusted R-squared is used (i.e. `fixest::r2()`
with `type="war2"`).
}
\examples{
plotR2Adj(sca_data = sca(y = "Salnty", x = "T_degC",
                         controls = c("ChlorA", "O2Sat"),
                         data = bottles, progressBar = TRUE,
                         parallel = FALSE),
                     title = "Adjusted R^2");
plotR2Adj(sca_data = sca(y="Salnty", x="T_degC",
                         controls = c("ChlorA*O2Sat"),
                         data = bottles, progressBar = FALSE,
                         parallel = FALSE),
                     showIndex = FALSE, plotVars = FALSE);
\donttest{
plotR2Adj(sca_data = sca(y = "Salnty", x = "T_degC",
                         controls = c("ChlorA*NO3uM", "O2Sat*NO3uM"),
                         data = bottles,
                         progressBar = TRUE, parallel = TRUE, workers = 2));
}
}
