% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scp}
\alias{scp}
\title{Takes in the output of `sca` and returns a list with the data frame and
labels to make a plot to visualize the controls included in each spec curve
model.}
\usage{
scp(spec_data)
}
\arguments{
\item{spec_data}{A data frame output by `sca`.}
}
\value{
A list containing a data frame, control coefficients, and control
        names.
}
\description{
Takes in the output of `sca` and returns a list with the data frame and
labels to make a plot to visualize the controls included in each spec curve
model.
}
\examples{
scp(sca(y="Salnty", x="T_degC", c("ChlorA", "O2Sat"), data=bottles,
    progressBar=TRUE, parallel=FALSE));
}
