% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{speccurvieR-deprecated}
\alias{speccurvieR-deprecated}
\alias{plotCurve}
\alias{plotVars}
\alias{plotRMSE}
\alias{plotR2Adj}
\alias{plotAIC}
\alias{plotDeviance}
\alias{plotControlDistributions}
\alias{controlExtractor}
\alias{unAsIs}
\title{Deprecated functions in speccurvieR}
\usage{
plotCurve(...)

plotVars(...)

plotRMSE(...)

plotR2Adj(...)

plotAIC(...)

plotDeviance(...)

plotControlDistributions(...)

controlExtractor(...)

unAsIs(...)
}
\arguments{
\item{...}{Arguments passed on to the replacement function.}
}
\value{
The value returned by the corresponding replacement function.
}
\description{
As of version 1.0.0 the package uses snake_case names throughout. These
camelCase aliases are retained for backward compatibility: they emit a
deprecation warning and forward to their snake_case replacements (translating
any deprecated camelCase argument names automatically). They are scheduled
for removal in version 2.0.0.
}
\keyword{internal}
