% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speccurvieR-package.R
\docType{package}
\name{speccurvieR-package}
\alias{speccurvieR}
\alias{speccurvieR-package}
\title{speccurvieR: Easy, Fast, and Pretty Specification Curve Analysis}
\description{
speccurvieR is an R package aimed at making specification curve analysis easy, fast, and pretty. It improves upon existing offerings with additional features and tidyverse integration. Users can easily visualize and evaluate how their models behave under different specifications with a high degree of customization.
}
\author{
\strong{Maintainer}: Zayne Sember \email{zsember@ucsd.edu}

}
\keyword{internal}
